Move audio unit name editor code to the info editor module.
